Moved key classes down into backbone
