+++ skin/src/main/java/org/researchstack/skin/ui/adapter/OnboardingPagerAdapter.java
@@ -10,10 +10,9 @@ import android.widget.TextView;
 
 import org.researchstack.backbone.ui.ViewVideoActivity;
 import org.researchstack.backbone.ui.views.LocalWebView;
-import org.researchstack.backbone.utils.ResUtils;
 import org.researchstack.backbone.utils.TextUtils;
 import org.researchstack.skin.R;
-import org.researchstack.skin.ResourceManager;
+import org.researchstack.backbone.ResourceManager;
 import org.researchstack.skin.model.StudyOverviewModel;
 
 import java.util.List;
