+++ skin/src/main/java/org/researchstack/skin/utils/ConsentFormUtils.java
@@ -10,7 +10,7 @@ import org.researchstack.backbone.utils.LogExt;
 import org.researchstack.backbone.utils.ObservableUtils;
 import org.researchstack.backbone.utils.ResUtils;
 import org.researchstack.skin.R;
-import org.researchstack.skin.ResourceManager;
+import org.researchstack.backbone.ResourceManager;
 
 import java.io.File;
 import java.io.FileOutputStream;
